Split module graph program checking
