+++ src/typechecker/program_module_graph.rs
@@ -1,0 +1,84 @@
+use super::*;
+
+impl TypeChecker {
+    pub fn check_module_graph_entry(
+        &mut self,
+        graph: &ResolvedModuleGraph,
+    ) -> Result<TypedProgram, Vec<Diagnostic>> {
+        let Some(entry) = graph.module(graph.entry) else {
+            self.diagnostics.push(Diagnostic::error(
+                "E0232",
+                format!("module graph missing entry module {:?}", graph.entry),
+                Span::dummy(),
+            ));
+            return Err(self
+                .diagnostics
+                .iter()
+                .filter(|diag| diag.is_error())
+                .cloned()
+                .collect());
+        };
+
+        let mut modules = graph.modules().values().collect::<Vec<_>>();
+        modules.sort_by_key(|module| module.info.id.0);
+
+        let mut dependency_programs = Vec::new();
+        for module in modules {
+            if module.info.id == graph.entry {
+                continue;
+            }
+
+            let mut checker = TypeChecker::new();
+            match checker.check_module_graph_module(graph, module) {
+                Ok(typed) => dependency_programs.push(typed),
+                Err(diags) => self.diagnostics.extend(diags),
+            }
+        }
+
+        if self.diagnostics.iter().any(|diag| diag.is_error()) {
+            return Err(self
+                .diagnostics
+                .iter()
+                .filter(|diag| diag.is_error())
+                .cloned()
+                .collect());
+        }
+
+        let mut typed = self.check_module_graph_module(graph, entry)?;
+        for mut dependency in dependency_programs {
+            typed.functions.append(&mut dependency.functions);
+            typed.types.append(&mut dependency.types);
+            typed.globals.append(&mut dependency.globals);
+        }
+        Ok(typed)
+    }
+
+    fn check_module_graph_module(
+        &mut self,
+        graph: &ResolvedModuleGraph,
+        module: &ResolvedModule,
+    ) -> Result<TypedProgram, Vec<Diagnostic>> {
+        self.validate_resolver_symbols(&module.program, &module.symbols);
+        if self.diagnostics.iter().any(|diag| diag.is_error()) {
+            return Err(self
+                .diagnostics
+                .iter()
+                .filter(|diag| diag.is_error())
+                .cloned()
+                .collect());
+        }
+
+        self.collect_module_graph_imports(graph, module);
+        if self.diagnostics.iter().any(|diag| diag.is_error()) {
+            return Err(self
+                .diagnostics
+                .iter()
+                .filter(|diag| diag.is_error())
+                .cloned()
+                .collect());
+        }
+
+        self.collect_declarations_with_symbols(&module.program.declarations, &module.symbols);
+        self.check_program_after_collection(&module.program)
+    }
+}
