+++ tests/docs_truth/repo_hygiene/mod.rs
@@ -18,6 +18,7 @@ mod resolver_validation;
 mod source_truth;
 mod typechecker_behavior_impls;
 mod typechecker_imports;
+mod typechecker_program_checking;
 mod typechecker_resolver_validation;
 mod typechecker_runtime;
 mod typechecker_semantic_validation;
